requirements.txt
Flask==1.0.2
urllib3==1.24.1
requests==2.21.0
IMAPClient==2.1.0
python_dateutil==2.7.5
python-dotenv==0.10.1
python-crontab==2.3.6
pytest==4.4.0
requests==2.21.0
gunicorn==19.9.0
flask-executor==0.9.2
beautifulsoup4==4.7.1
locustio==0.11.0
cron_descriptor==1.2.21